[FTR] split x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/config.ts
