+++ .buildkite/ftr_platform_stateful_configs.yml
@@ -252,7 +252,9 @@ enabled:
   - x-pack/platform/test/functional/apps/maps/group2/config.ts
   - x-pack/platform/test/functional/apps/maps/group3/config.ts
   - x-pack/platform/test/functional/apps/maps/group4/config.ts
-  - x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/config.ts
+  - x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group1/config.ts
+  - x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group2/config.ts
+  - x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group3/config.ts
   - x-pack/platform/test/functional/apps/ml/anomaly_detection_integrations/config.ts
   - x-pack/platform/test/functional/apps/ml/anomaly_detection_result_views/config.ts
   - x-pack/platform/test/functional/apps/ml/data_frame_analytics/group1/config.ts
