+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group1/index.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService, loadTestFile }: FtrProviderContext) {
   const config = getService('config');
@@ -13,7 +13,7 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
   const esNode = isCcs ? getService('remoteEsArchiver' as 'esArchiver') : getService('esArchiver');
   const ml = getService('ml');
 
-  describe('machine learning - anomaly detection', function () {
+  describe('machine learning - anomaly detection - group1', function () {
     this.tags(['skipFirefox']);
 
     before(async () => {
@@ -30,9 +30,6 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
       await ml.securityCommon.cleanMlRoles();
 
       await esNode.unload('x-pack/platform/test/fixtures/es_archives/ml/farequote');
-      await esNode.unload('x-pack/platform/test/fixtures/es_archives/ml/ecommerce');
-      await esNode.unload('x-pack/platform/test/fixtures/es_archives/ml/categorization_small');
-      await esNode.unload('x-pack/platform/test/fixtures/es_archives/ml/event_rate_nanos');
 
       await ml.testResources.resetKibanaTimeZone();
     });
@@ -43,17 +40,7 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
       loadTestFile(require.resolve('./job_expanded_details'));
       loadTestFile(require.resolve('./single_metric_job_without_datafeed_start'));
       loadTestFile(require.resolve('./multi_metric_job'));
-      loadTestFile(require.resolve('./population_job'));
-      loadTestFile(require.resolve('./geo_job'));
       loadTestFile(require.resolve('./saved_search_job'));
-      loadTestFile(require.resolve('./advanced_job'));
-      loadTestFile(require.resolve('./categorization_job'));
-      loadTestFile(require.resolve('./date_nanos_job'));
-      loadTestFile(require.resolve('./custom_urls'));
-      loadTestFile(require.resolve('./delete_job_and_delete_annotations'));
-      loadTestFile(require.resolve('./convert_single_metric_job_to_multi_metric'));
-      loadTestFile(require.resolve('./convert_jobs_to_advanced_job'));
-      loadTestFile(require.resolve('./supplied_configurations'));
     }
   });
 }
