+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group1/job_expanded_details.ts
@@ -7,8 +7,8 @@
 
 import type { Job } from '@kbn/ml-plugin/common/types/anomaly_detection_jobs';
 import { JOB_STATE } from '@kbn/ml-plugin/common';
-import { QuickFilterButtonTypes } from '../../../services/ml/job_table';
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+import { QuickFilterButtonTypes } from '../../../../services/ml/job_table';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
