+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group1/single_metric_job_without_datafeed_start.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
