+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group2/config.ts
@@ -8,13 +8,13 @@
 import type { FtrConfigProviderContext } from '@kbn/test';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
-  const functionalConfig = await readConfigFile(require.resolve('../../../config.base.ts'));
+  const baseConfig = await readConfigFile(require.resolve('../group1/config.ts'));
 
   return {
-    ...functionalConfig.getAll(),
+    ...baseConfig.getAll(),
     testFiles: [require.resolve('.')],
     junit: {
-      reportName: 'Chrome X-Pack UI Functional Tests - ML anomaly_detection_jobs',
+      reportName: 'Chrome X-Pack UI Functional Tests - ML anomaly_detection_jobs - group2',
     },
   };
 }
