+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group2/geo_job.ts
@@ -5,57 +5,62 @@
  * 2.0.
  */
 
-import { CATEGORY_EXAMPLES_VALIDATION_STATUS } from '@kbn/ml-category-validator';
-import type { FtrProviderContext } from '../../../ftr_provider_context';
-import type { FieldStatsType } from '../common/types';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
   const ml = getService('ml');
 
-  const jobId = `categorization_${Date.now()}`;
+  const jobId = `ec_geo_1_${Date.now()}`;
   const jobIdClone = `${jobId}_clone`;
-  const jobDescription =
-    'Create categorization job based on the ft_categorization dataset with a count rare';
-  const jobGroups = ['automated', 'categorization'];
+  const jobDescription = 'Create geo job based on the ecommerce sample dataset with 15m bucketspan';
+  const jobGroups = ['automated', 'ecommerce', 'geo'];
   const jobGroupsClone = [...jobGroups, 'clone'];
-  const detectorTypeIdentifier = 'Rare';
-  const categorizationFieldIdentifier = 'field1';
-  const categorizationExampleCount = 5;
-  const bucketSpan = '1d';
-  const memoryLimit = '15mb';
+  const geoField = 'geoip.location';
+  const splitField = 'customer_gender';
+  const detectors = [
+    {
+      identifier: 'Lat long(geoip.location)',
+      frontCardTitle: "Men's Clothing",
+      function: 'lat_long',
+      field_name: geoField,
+      numberOfBackCards: 5,
+      splitField,
+    },
+  ];
+  const bucketSpan = '15m';
+  const memoryLimit = '8mb';
 
   function getExpectedRow(expectedJobId: string, expectedJobGroups: string[]) {
     return {
       id: expectedJobId,
       description: jobDescription,
       jobGroups: [...new Set(expectedJobGroups)].sort(),
-      recordCount: '1,000',
+      recordCount: '4,675',
       memoryStatus: 'ok',
       jobState: 'closed',
       datafeedState: 'stopped',
-      latestTimestamp: '2019-11-21 00:01:13',
+      latestTimestamp: '2023-07-12 23:45:36',
     };
   }
 
   function getExpectedCounts(expectedJobId: string) {
     return {
       job_id: expectedJobId,
-      processed_record_count: '1,000',
-      processed_field_count: '1,000',
-      input_bytes: '148.8 KB',
-      input_field_count: '1,000',
+      processed_record_count: '4,675',
+      processed_field_count: '9,350',
+      input_bytes: '504.1 KB',
+      input_field_count: '9,350',
       invalid_date_count: '0',
       missing_field_count: '0',
       out_of_order_timestamp_count: '0',
-      empty_bucket_count: '23',
+      empty_bucket_count: '492',
       sparse_bucket_count: '0',
-      bucket_count: '230',
-      earliest_record_timestamp: '2019-04-05 11:25:35',
-      latest_record_timestamp: '2019-11-21 00:01:13',
-      input_record_count: '1,000',
-      latest_bucket_timestamp: '2019-11-21 00:00:00',
-      latest_empty_bucket_timestamp: '2019-11-17 00:00:00',
+      bucket_count: '2,975',
+      earliest_record_timestamp: '2023-06-12 00:04:19',
+      latest_record_timestamp: '2023-07-12 23:45:36',
+      input_record_count: '4,675',
+      latest_bucket_timestamp: '2023-07-12 23:45:00',
     };
   }
 
@@ -64,31 +69,22 @@ export default function ({ getService }: FtrProviderContext) {
       job_id: expectedJobId,
       result_type: 'model_size_stats',
       model_bytes_exceeded: '0.0 B',
-      // not checking total_by_field_count as the number of categories might change
+      total_by_field_count: '4',
       total_over_field_count: '0',
-      total_partition_field_count: '2',
+      total_partition_field_count: '3',
       bucket_allocation_failures_count: '0',
       memory_status: 'ok',
-      timestamp: '2019-11-20 00:00:00',
+      timestamp: '2023-07-12 23:30:00',
     };
   }
 
   const calendarId = `wizard-test-calendar_${Date.now()}`;
 
-  const fieldStatsEntries = [
-    {
-      fieldName: 'field1',
-      type: 'keyword' as FieldStatsType,
-    },
-  ];
-
-  describe('categorization', function () {
+  describe('geo', function () {
     this.tags(['ml']);
     before(async () => {
-      await esArchiver.loadIfNeeded(
-        'x-pack/platform/test/fixtures/es_archives/ml/categorization_small'
-      );
-      await ml.testResources.createDataViewIfNeeded('ft_categorization_small', '@timestamp');
+      await esArchiver.loadIfNeeded('x-pack/platform/test/fixtures/es_archives/ml/ecommerce');
+      await ml.testResources.createDataViewIfNeeded('ft_ecommerce', 'order_date');
       await ml.testResources.setKibanaTimeZoneToUTC();
 
       await ml.api.createCalendar(calendarId);
@@ -97,33 +93,31 @@ export default function ({ getService }: FtrProviderContext) {
 
     after(async () => {
       await ml.api.cleanMlIndices();
-      await ml.testResources.deleteDataViewByTitle('ft_categorization_small');
+      await ml.testResources.deleteDataViewByTitle('ft_ecommerce');
     });
 
-    it('job creation loads the categorization wizard for the source data', async () => {
+    it('job creation loads the geo wizard for the source data', async () => {
       await ml.testExecution.logTestStep('job creation loads the job management page');
-      await ml.testExecution.logTestStep('');
       await ml.navigation.navigateToStackManagementMlSection('anomaly_detection', 'ml-jobs-list');
 
       await ml.testExecution.logTestStep('job creation loads the new job source selection page');
       await ml.jobManagement.navigateToNewJobSourceSelection();
 
       await ml.testExecution.logTestStep('job creation loads the job type selection page');
-      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_categorization_small');
+      await ml.jobSourceSelection.selectSourceForAnomalyDetectionJob('ft_ecommerce');
 
-      await ml.testExecution.logTestStep('job creation loads the categorization job wizard page');
-      await ml.jobTypeSelection.selectCategorizationJob();
+      await ml.testExecution.logTestStep('job creation loads the geo job wizard page');
+      await ml.jobTypeSelection.selectGeoJob();
     });
 
-    it('job creation navigates through the categorization wizard and sets all needed fields', async () => {
+    it('job creation navigates through the geo wizard and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job creation displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
-      await ml.commonUI.assertDatePickerDataTierOptionsVisible(true);
 
       await ml.testExecution.logTestStep('job creation sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Apr 5, 2019 @ 11:25:35.770',
-        'Nov 21, 2019 @ 00:01:13.923'
+        'Jun 12, 2023 @ 00:04:19.000',
+        'Jul 12, 2023 @ 23:45:36.000'
       );
 
       await ml.testExecution.logTestStep('job creation displays the event rate chart');
@@ -133,31 +127,28 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job creation displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
-      await ml.testExecution.logTestStep(
-        `job creation selects ${detectorTypeIdentifier} detector type`
-      );
-      await ml.jobWizardCategorization.assertCategorizationDetectorTypeSelectionExists();
-      await ml.jobWizardCategorization.selectCategorizationDetectorType(detectorTypeIdentifier);
+      await ml.testExecution.logTestStep('job creation selects the geo field');
+      await ml.jobWizardGeo.assertGeoFieldInputExists();
+      await ml.jobWizardGeo.selectGeoField(geoField);
+
+      await ml.testExecution.logTestStep('job creation displays detector preview');
+
+      await ml.jobWizardGeo.assertDetectorPreviewExists(detectors[0].identifier);
 
       await ml.testExecution.logTestStep(
-        'job creation opens field stats flyout from categorization field input'
+        'job creation inputs the split field and displays split cards'
       );
-      await ml.jobWizardCategorization.assertCategorizationFieldInputExists();
-      for (const { fieldName, type: fieldType } of fieldStatsEntries) {
-        await ml.jobWizardCategorization.assertFieldStatFlyoutContentFromCategorizationFieldInputTrigger(
-          fieldName,
-          fieldType
-        );
-      }
 
-      await ml.testExecution.logTestStep(`job creation selects the categorization field`);
-      await ml.jobWizardCategorization.selectCategorizationField(categorizationFieldIdentifier);
-      await ml.jobWizardCategorization.assertCategorizationExamplesCallout(
-        CATEGORY_EXAMPLES_VALIDATION_STATUS.VALID
-      );
-      await ml.jobWizardCategorization.assertCategorizationExamplesTable(
-        categorizationExampleCount
-      );
+      await ml.jobWizardMultiMetric.assertSplitFieldInputExists();
+      await ml.jobWizardMultiMetric.selectSplitField(splitField);
+
+      await ml.jobWizardMultiMetric.assertDetectorSplitExists(splitField);
+      await ml.jobWizardMultiMetric.assertDetectorSplitFrontCardTitle('FEMALE');
+      await ml.jobWizardMultiMetric.assertDetectorSplitNumberOfBackCards(1);
+
+      await ml.testExecution.logTestStep('job creation displays the influencer field');
+      await ml.jobWizardCommon.assertInfluencerInputExists();
+      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
 
       await ml.testExecution.logTestStep('job creation inputs the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -195,8 +186,6 @@ export default function ({ getService }: FtrProviderContext) {
 
       await ml.testExecution.logTestStep('job creation displays the model plot switch');
       await ml.jobWizardCommon.assertModelPlotSwitchExists();
-      await ml.jobWizardCommon.assertModelPlotSwitchEnabled(false);
-      await ml.jobWizardCommon.assertModelPlotSwitchCheckedState(false);
 
       await ml.testExecution.logTestStep('job creation enables the dedicated index switch');
       await ml.jobWizardCommon.assertDedicatedIndexSwitchExists();
@@ -235,25 +224,27 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job creation has detector results');
-      await ml.api.assertDetectorResultsExist(jobId, 0);
+      for (let i = 0; i < detectors.length; i++) {
+        await ml.api.assertDetectorResultsExist(jobId, i);
+      }
     });
 
-    it('job cloning opens the existing job in the categorization wizard', async () => {
+    it('job cloning opens the existing job in the geo wizard', async () => {
       await ml.testExecution.logTestStep(
-        'job cloning clicks the clone action and loads the single metric wizard'
+        'job cloning clicks the clone action and loads the geo wizard'
       );
       await ml.jobTable.clickCloneJobAction(jobId);
-      await ml.jobTypeSelection.assertCategorizationJobWizardOpen();
+      await ml.jobTypeSelection.assertGeoJobWizardOpen();
     });
 
-    it('job cloning navigates through the categorization wizard, checks and sets all needed fields', async () => {
+    it('job cloning navigates through the geo wizard, checks and sets all needed fields', async () => {
       await ml.testExecution.logTestStep('job cloning displays the time range step');
       await ml.jobWizardCommon.assertTimeRangeSectionExists();
 
       await ml.testExecution.logTestStep('job cloning sets the time range');
       await ml.jobWizardCommon.clickUseFullDataButton(
-        'Apr 5, 2019 @ 11:25:35.770',
-        'Nov 21, 2019 @ 00:01:13.923'
+        'Jun 12, 2023 @ 00:04:19.000',
+        'Jul 12, 2023 @ 23:45:36.000'
       );
 
       await ml.testExecution.logTestStep('job cloning displays the event rate chart');
@@ -263,8 +254,18 @@ export default function ({ getService }: FtrProviderContext) {
       await ml.testExecution.logTestStep('job cloning displays the pick fields step');
       await ml.jobWizardCommon.advanceToPickFieldsSection();
 
-      await ml.testExecution.logTestStep('job cloning pre-fills field and aggregation');
-      await ml.jobWizardCategorization.assertCategorizationDetectorTypeSelectionExists();
+      await ml.testExecution.logTestStep('job cloning pre-fills the geo field');
+      await ml.jobWizardGeo.assertGeoFieldInputExists();
+      await ml.jobWizardGeo.assertGeoFieldSelection([geoField]);
+
+      await ml.testExecution.logTestStep(
+        'job cloning displays the detector preview with pre-filled geo field'
+      );
+      await ml.jobWizardGeo.assertDetectorPreviewExists(detectors[0].identifier);
+
+      await ml.testExecution.logTestStep('job cloning pre-fills influencers');
+      await ml.jobWizardCommon.assertInfluencerInputExists();
+      await ml.jobWizardCommon.assertInfluencerSelection([splitField]);
 
       await ml.testExecution.logTestStep('job cloning pre-fills the bucket span');
       await ml.jobWizardCommon.assertBucketSpanInputExists();
@@ -307,18 +308,11 @@ export default function ({ getService }: FtrProviderContext) {
 
       await ml.testExecution.logTestStep('job cloning pre-fills the model plot switch');
       await ml.jobWizardCommon.assertModelPlotSwitchExists();
-      await ml.jobWizardCommon.assertModelPlotSwitchEnabled(false);
       await ml.jobWizardCommon.assertModelPlotSwitchCheckedState(false);
 
       await ml.testExecution.logTestStep('job cloning pre-fills the dedicated index switch');
       await ml.jobWizardCommon.assertDedicatedIndexSwitchExists();
       await ml.jobWizardCommon.assertDedicatedIndexSwitchCheckedState(true);
-
-      // MML during clone has changed in #61589
-      // TODO: adjust test code to reflect the new behavior
-      // await ml.testExecution.logTestStep('job cloning pre-fills the model memory limit');
-      // await ml.jobWizardCommon.assertModelMemoryLimitInputExists();
-      // await ml.jobWizardCommon.assertModelMemoryLimitValue(memoryLimit);
 
       await ml.testExecution.logTestStep('job cloning displays the validation step');
       await ml.jobWizardCommon.advanceToValidationSection();
@@ -349,29 +343,9 @@ export default function ({ getService }: FtrProviderContext) {
       );
 
       await ml.testExecution.logTestStep('job cloning has detector results');
-      await ml.api.assertDetectorResultsExist(jobId, 0);
-    });
-
-    it('deletes the cloned job', async () => {
-      await ml.testExecution.logTestStep('job deletion has results for the job before deletion');
-      await ml.api.assertJobResultsExist(jobIdClone);
-
-      await ml.testExecution.logTestStep('job deletion triggers the delete action');
-      await ml.jobTable.clickDeleteJobAction(jobIdClone);
-
-      await ml.testExecution.logTestStep('job deletion confirms the delete modal');
-      await ml.jobTable.confirmDeleteJobModal();
-      await ml.api.waitForAnomalyDetectionJobNotToExist(jobIdClone, 30 * 1000);
-
-      await ml.testExecution.logTestStep(
-        'job deletion does not display the deleted job in the job list any more'
-      );
-      await ml.jobTable.filterWithSearchString(jobIdClone, 0);
-
-      await ml.testExecution.logTestStep(
-        'job deletion does not have results for the deleted job any more'
-      );
-      await ml.api.assertNoJobResultsExist(jobIdClone);
+      for (let i = 0; i < detectors.length; i++) {
+        await ml.api.assertDetectorResultsExist(jobId, i);
+      }
     });
   });
 }
