+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group2/index.ts
@@ -1,0 +1,44 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
+
+export default function ({ getService, loadTestFile }: FtrProviderContext) {
+  const esArchiver = getService('esArchiver');
+  const ml = getService('ml');
+
+  describe('machine learning - anomaly detection - group2', function () {
+    this.tags(['skipFirefox']);
+
+    before(async () => {
+      await ml.securityCommon.createMlRoles();
+      await ml.securityCommon.createMlUsers();
+      await ml.securityUI.loginAsMlPowerUser();
+    });
+
+    after(async () => {
+      // NOTE: Logout needs to happen before anything else to avoid flaky behavior
+      await ml.securityUI.logout();
+
+      await ml.securityCommon.cleanMlUsers();
+      await ml.securityCommon.cleanMlRoles();
+
+      await esArchiver.unload('x-pack/platform/test/fixtures/es_archives/ml/ecommerce');
+      await esArchiver.unload('x-pack/platform/test/fixtures/es_archives/ml/categorization_small');
+      await esArchiver.unload('x-pack/platform/test/fixtures/es_archives/ml/event_rate_nanos');
+      await esArchiver.unload('x-pack/platform/test/fixtures/es_archives/ml/farequote');
+
+      await ml.testResources.resetKibanaTimeZone();
+    });
+
+    loadTestFile(require.resolve('./population_job'));
+    loadTestFile(require.resolve('./geo_job'));
+    loadTestFile(require.resolve('./categorization_job'));
+    loadTestFile(require.resolve('./date_nanos_job'));
+    loadTestFile(require.resolve('./custom_urls'));
+  });
+}
