+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group3/advanced_job.ts
@@ -5,8 +5,8 @@
  * 2.0.
  */
 
-import type { FtrProviderContext } from '../../../ftr_provider_context';
-import type { PickFieldsConfig, DatafeedConfig, Detector } from './types';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
+import type { PickFieldsConfig, DatafeedConfig, Detector } from '../types';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
