+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group3/convert_jobs_to_advanced_job.ts
@@ -6,8 +6,8 @@
  */
 
 import { CATEGORY_EXAMPLES_VALIDATION_STATUS } from '@kbn/ml-category-validator';
-import type { PickFieldsConfig, DatafeedConfig, Detector } from './types';
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+import type { PickFieldsConfig, DatafeedConfig, Detector } from '../types';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
