+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group3/convert_single_metric_job_to_multi_metric.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const config = getService('config');
