+++ x-pack/platform/test/functional/apps/ml/anomaly_detection_jobs/group3/delete_job_and_delete_annotations.ts
@@ -9,8 +9,8 @@ import { ANNOTATION_TYPE } from '@kbn/ml-plugin/common/constants/annotations';
 import {
   SINGLE_METRIC_JOB_CONFIG,
   MULTI_METRIC_JOB_CONFIG,
-} from '../../../../api_integration/apis/ml/jobs/common_jobs';
-import type { FtrProviderContext } from '../../../ftr_provider_context';
+} from '../../../../../api_integration/apis/ml/jobs/common_jobs';
+import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ getService }: FtrProviderContext) {
   const config = getService('config');
